requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
python-socketio==5.10.0
geopy==2.4.1
pydantic==2.5.0
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.2
pandas==2.1.3
python-dotenv==1.0.0
